modules/person-service/requirements.txt
grpcio
grpcio-tools
SQLAlchemy==1.3.19
Werkzeug==0.16.1
Flask-SQLAlchemy==2.4.0
GeoAlchemy2==0.8.4